Rename testgallery module directory and source package to e2e-mock-gallery/mockgallery
Moves the module directory from testgallery/ to e2e-mock-gallery/ and relocates
the Kotlin source from com.gb4pc.testgallery to com.gb4pc.mockgallery, to match
the naming convention established by e2e-mock-camera.
