+++ e2e-mock-gallery/build.gradle.kts
@@ -6,7 +6,7 @@ plugins {
 /**
  * Minimal test-only APK used by the E2E visual test suite.
  *
- * applicationId = "com.gb4pc.testgallery" is the gallery package seeded into
+ * applicationId = "com.gb4pc.mockgallery" is the gallery package seeded into
  * PrefsManager by E2EFixture.seedGalleryPrefs() so that GB4PC's overlay points
  * to this app during visual tests.
  *
@@ -14,11 +14,11 @@ plugins {
  * giving the tests a real gallery-app target without requiring a shipping gallery.
  */
 android {
-    namespace = "com.gb4pc.testgallery"
+    namespace = "com.gb4pc.mockgallery"
     compileSdk = 35
 
     defaultConfig {
-        applicationId = "com.gb4pc.testgallery"
+        applicationId = "com.gb4pc.mockgallery"
         minSdk = 26
         targetSdk = 35
         versionCode = 1
