+++ e2e-mock-gallery/src/main/AndroidManifest.xml
