+++ e2e-mock-gallery/src/main/res/drawable/ic_launcher_background.xml
