+++ e2e-mock-gallery/src/main/res/layout/activity_last_photo.xml
