+++ e2e-mock-gallery/src/main/res/mipmap-anydpi-v26/ic_launcher.xml
